+++ packages/limps/CHANGELOG.md
@@ -20,8 +20,8 @@
 
 ### Features
 
-* **limps-radix,claude:** implement agents 007-008 and add review skills ([#39](https://github.com/paulbreuler/limps/issues/39)) ([777e107](https://github.com/paulbreuler/limps/commit/777e10765b10ed64c4b9ddede4d566696b1f0ab2))
-* **limps-radix:** implement Radix UI extension with signature generation and caching ([#36](https://github.com/paulbreuler/limps/issues/36)) ([1134be0](https://github.com/paulbreuler/limps/commit/1134be033aab412fdffa911f929512a6811038dc))
+* **limps-headless,claude:** implement agents 007-008 and add review skills ([#39](https://github.com/paulbreuler/limps/issues/39)) ([777e107](https://github.com/paulbreuler/limps/commit/777e10765b10ed64c4b9ddede4d566696b1f0ab2))
+* **limps-headless:** implement Radix UI extension with signature generation and caching ([#36](https://github.com/paulbreuler/limps/issues/36)) ([1134be0](https://github.com/paulbreuler/limps/commit/1134be033aab412fdffa911f929512a6811038dc))
 
 ## [2.5.0](https://github.com/paulbreuler/limps/compare/v2.4.0...v2.5.0) (2026-01-27)